pyproject.toml
[project]
name = "biohacker-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
